refactor: setting モジュール構造を簡略化
