+++ build_util/make_docs.py
@@ -7,7 +7,7 @@ from voicevox_engine.core.core_initializer import CoreManager
 from voicevox_engine.dev.core.mock import MockCoreWrapper
 from voicevox_engine.dev.tts_engine.mock import MockTTSEngine
 from voicevox_engine.preset.PresetManager import PresetManager
-from voicevox_engine.setting.Setting import USER_SETTING_PATH, SettingHandler
+from voicevox_engine.setting import USER_SETTING_PATH, SettingHandler
 from voicevox_engine.tts_pipeline.tts_engine import TTSEngineManager
 from voicevox_engine.user_dict.user_dict import UserDictionary
 from voicevox_engine.utility.path_utility import engine_root
