+++ run.py
@@ -15,11 +15,7 @@ from voicevox_engine.app.application import generate_app
 from voicevox_engine.cancellable_engine import CancellableEngine
 from voicevox_engine.core.core_initializer import initialize_cores
 from voicevox_engine.preset.PresetManager import PresetManager
-from voicevox_engine.setting.Setting import (
-    USER_SETTING_PATH,
-    CorsPolicyMode,
-    SettingHandler,
-)
+from voicevox_engine.setting import USER_SETTING_PATH, CorsPolicyMode, SettingHandler
 from voicevox_engine.tts_pipeline.tts_engine import make_tts_engines_from_cores
 from voicevox_engine.user_dict.user_dict import UserDictionary
 from voicevox_engine.utility.path_utility import engine_root
