+++ voicevox_engine/app/routers/setting.py
@@ -6,7 +6,7 @@ from fastapi import APIRouter, Depends, Form, Request, Response
 from fastapi.templating import Jinja2Templates
 
 from voicevox_engine.engine_manifest.EngineManifest import EngineManifest
-from voicevox_engine.setting.Setting import CorsPolicyMode, Setting, SettingHandler
+from voicevox_engine.setting import CorsPolicyMode, Setting, SettingHandler
 from voicevox_engine.utility.path_utility import engine_root
 
 from ..dependencies import check_disabled_mutable_api
